refactor: move gitStatusLetter to domain module
Breaks false coupling between GitStatusBadge and the DnD composable.
gitStatusLetter is a pure domain mapping function that belongs in
src/lib/domain/git.ts, not use-item-interactions.svelte.ts.
