+++ src/lib/composables/use-item-interactions.svelte.ts
@@ -88,17 +88,3 @@ export function isInClipboard(entry: FileEntry): boolean {
 export function isClipboardCut(entry: FileEntry): boolean {
   return isInClipboard(entry) && clipboardStore.isCut;
 }
-
-// --- Git status helper ---
-
-/** Convert a git status string to its single-letter indicator */
-export function gitStatusLetter(status: string): string {
-  switch (status) {
-    case "Modified": return "M";
-    case "Untracked": return "U";
-    case "Added": return "A";
-    case "Deleted": return "D";
-    case "Conflict": return "!";
-    default: return "R";
-  }
-}
